Safari-611.1.10/LayoutTests/editing/spelling/spelling-double-clicked-word-expected.txt

Spelling markers should remain for double-clicked misspellings. To test manually, type 'wellcome home.', double-click on 'wellcome'. The test succeeds when 'wellcome' remains spelling marker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.markerCountForNode(destination.childNodes[0], "spelling") became different from 0
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome"
PASS spellingMarkerRange.toString() became "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome "
PASS spellingMarkerRange.toString() became "wellcome"
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
